Add profile management features: implement commands and requests for saving user profile, skills, education, experience, and attachments; update models and handlers for new data structures; enhance UI components for profile steps.

+++ src/Infrastructure/Services/StorageServices/LocalStorageService.cs
@@ -50,7 +50,7 @@ public sealed class LocalStorageService : IFileStorageService
             File.Move(tmp, full);
 
             var size = new FileInfo(full).Length;
-            return Result.Ok(new FileSaved(blobKey, size));
+            return Result.Ok(new FileSaved(blobKey, (ulong)size));
         }
         catch (OperationCanceledException oce)
         {
